+++ src/main/java/org/sonatype/nexus/repository/chef/internal/AssetKind.java
@@ -10,22 +10,36 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.repository.chef.internal;
+package org.sonatype.nexus.repository.chef.internal;
 
-import org.junit.Test;
+import javax.annotation.Nonnull;
 
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
+import org.sonatype.nexus.repository.cache.CacheControllerHolder.CacheType;
 
-public class ChefFormatTest
+import static org.sonatype.nexus.repository.cache.CacheControllerHolder.*;
+
+/**
+ * Asset kinds for Chef
+ *
+ * Examples of all can be found at: https://github.com/kubernetes/helm/blob/master/docs/chart_repository.md#the-chart-repository-structure
+ * @since 0.0.1
+ */
+public enum AssetKind
 {
-  private ChefFormat underTest;
+  COOKBOOK_DETAIL_VERSION(METADATA),
+  COOKBOOK_DETAILS(METADATA),
+  COOKBOOKS_LIST(METADATA),
+  COOKBOOKS_SEARCH(METADATA),
+  COOKBOOK(CONTENT);
 
-  @Test
-  public void checkFormatNameIsCorrect() {
-    underTest = new ChefFormat();
+  private final CacheType cacheType;
 
-    assertThat(underTest.getValue(), is(equalTo("chef")));
+  AssetKind(final CacheType cacheType) {
+    this.cacheType = cacheType;
+  }
+
+  @Nonnull
+  public CacheType getCacheType() {
+    return cacheType;
   }
 }
