+++ src/main/java/org/sonatype/nexus/repository/chef/internal/metadata/ChefAttributes.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.repository.chef.internal.metadata;
+package org.sonatype.nexus.repository.chef.internal.metadata;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
