+++ src/main/java/org/sonatype/nexus/repository/chef/internal/security/ChefFormatSecurityContributor.java
@@ -10,26 +10,29 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.repository.chef.internal;
+package org.sonatype.nexus.repository.chef.internal.security;
 
+import javax.inject.Inject;
 import javax.inject.Named;
 import javax.inject.Singleton;
 
+import org.sonatype.nexus.repository.chef.internal.ChefFormat;
+
 import org.sonatype.nexus.repository.Format;
+import org.sonatype.nexus.repository.security.RepositoryFormatSecurityContributor;
 
 /**
- * Chef repository format.
+ * Chef format security resource.
  *
  * @since 0.0.1
  */
-@Named(ChefFormat.NAME)
+@Named
 @Singleton
-public class ChefFormat
-    extends Format
+public class ChefFormatSecurityContributor
+    extends RepositoryFormatSecurityContributor
 {
-  public static final String NAME = "chef";
-
-  public ChefFormat() {
-    super(NAME);
+  @Inject
+  public ChefFormatSecurityContributor(@Named(ChefFormat.NAME) final Format format) {
+    super(format);
   }
 }
