+++ src/main/java/org/sonatype/nexus/repository/chef/internal/security/ChefSecurityFacet.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.repository.chef.internal.security;
+package org.sonatype.nexus.repository.chef.internal.security;
 
 import javax.inject.Inject;
 import javax.inject.Named;
