+++ src/test/java/org/sonatype/nexus/repository/chef/internal/ChefFormatTest.java
@@ -10,29 +10,22 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.repository.chef.internal.security;
+package org.sonatype.nexus.repository.chef.internal;
 
-import javax.inject.Inject;
-import javax.inject.Named;
-import javax.inject.Singleton;
+import org.junit.Test;
 
-import org.sonatype.repository.chef.internal.ChefFormat;
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
 
-import org.sonatype.nexus.repository.Format;
-import org.sonatype.nexus.repository.security.RepositoryFormatSecurityContributor;
-
-/**
- * Chef format security resource.
- *
- * @since 0.0.1
- */
-@Named
-@Singleton
-public class ChefFormatSecurityContributor
-    extends RepositoryFormatSecurityContributor
+public class ChefFormatTest
 {
-  @Inject
-  public ChefFormatSecurityContributor(@Named(ChefFormat.NAME) final Format format) {
-    super(format);
+  private ChefFormat underTest;
+
+  @Test
+  public void checkFormatNameIsCorrect() {
+    underTest = new ChefFormat();
+
+    assertThat(underTest.getValue(), is(equalTo("chef")));
   }
 }
